requirements.txt
wheel
aiohttp[speedups]>=3.5.4
aiofiles
aria2p
async_generator
beautifulsoup4>=4.8.1
cowpy>=1.1.0
psutil
wget
telegraph
TgCrypto>=1.2.2
cffi>=1.0.0
dnspython
emoji
PyPDF2
cryptg
patool
gitpython
asyncurban>=0.3.3
google-api-python-client==2.9.0
google-auth-oauthlib
googletrans==4.0.0-rc1
google_images_download>=2.7.1
gTTS>=2.1.1
gTTS-token>=1.1.3
hachoir>=3.0a5
REPO_LINK
heroku3>=3.4.0
httplib2>=0.18.0
humanize
meval>=2.5
lxml
lyrics_extractor
oauth2client
Pillow>=8.1.0
fontTools
psycopg2
psycopg2-binary
pybase64>=0.4.0
pylast
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
requests>=2.25.1
search-engine-parser>=0.6.1
speedtest-cli>=2.1.2
sqlalchemy<=1.3.23
telethon>=1.14.0
telethon-tgcrypto
telethon-session-sqlalchemy>=0.2.16
urbandict>=0.5
wikipedia>=1.4.0
coffeehouse>=2.1.0
gitpython
git+https://github.com/TomyPrs/deezloader
git+https://github.com/johnwmillr/LyricsGenius.git
git+https://github.com/voins/google-images-download.git@2cd68173c961324a2c41c61e8b6f40a49663ce60
lyricsgenius
redis>=3.3.10
redis_db
PyDrive2
pyDownload
beautifulsoup4>=4.8.1
PyGithub
bs4
covid>=2.2.9
base
pip>=19.2.3
setuptools>=28.8.0
selenium>=3.141.0
python-dateutil
pyfiglet
jotquote
html_telegraph_poster
csvfaker
git+https://github.com/LonamiWebs/Telethon
git+https://github.com/ahupp/python-magic
pytube>=9.5.1
pymongo[srv]
youtube-dl>=2021.1.8
youtube-search-python>=1.3.2
youtube-search>=1.1.1
justwatch
zipfile38
bitlyshortener
validators
geopy
glitch_this
twitter_scraper
pyaztro
jikanpy
vcsi
lottie
wand
colour
scipy
wordcloud
opencv-python
faker
pendulum
rarfile
natsort
hentai
Faker==8.8.0
